requirements.txt
python-dotenv
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
livekit-plugins-silero